Using unsigned long long everywhere

+++ omp_parallel_for_pi.C
@@ -32,10 +32,11 @@ int main()
     #pragma omp parallel
     {
         n_threads = omp_get_num_threads();
+
         // OpenMP can handle for us for loop ranges!
         // TIP: you have to add something in the following line...
 
-        for (long i=1; i<=common::num_steps; i++) {
+        for (unsigned long long i=1; i<=common::num_steps; i++) {
             double x = (i - 0.5) * step; //computing the x value
             sum += 4.0 / (1.0 + x * x); //adding to the cumulus
         }
